Small plugin structure fix
